moved pbar class
